Reorganizadas todas las clases en packages
Para mejor organización y entendimiento :)
https://i.imgur.com/7ttdvFR.png
