+++ src/colecciones/Administrador.java
@@ -1,3 +1,4 @@
+package colecciones;
 
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
