+++ src/colecciones/Cliente.java
@@ -1,3 +1,4 @@
+package colecciones;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
