+++ src/colecciones/Compania.java
@@ -1,3 +1,4 @@
+package colecciones;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -16,6 +17,8 @@ import com.itextpdf.text.Document;
 import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Paragraph;
 import com.itextpdf.text.pdf.PdfWriter;
+
+import extras.Database;
 
 /**
  * @author FValerio, DMayorga, MSilva, LMondaca
