+++ src/colecciones/Equipo.java
@@ -1,3 +1,4 @@
+package colecciones;
 /**
  *
  */
