+++ src/colecciones/Persona.java
@@ -1,3 +1,4 @@
+package colecciones;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.util.ArrayList;
