+++ src/colecciones/RegistroDePagos.java
@@ -1,3 +1,4 @@
+package colecciones;
 public class RegistroDePagos{
 	private int idContrato;
 	private int idRegistro;
