+++ src/extras/Database.java
@@ -1,9 +1,18 @@
+package extras;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+
+import colecciones.Administrador;
+import colecciones.Cliente;
+import colecciones.Compania;
+import colecciones.Contrato;
+import colecciones.Equipo;
+import colecciones.Plan;
+import colecciones.RegistroDePagos;
 
 /**
  *
