+++ src/interfaz/FrameInterfaz.java
@@ -1,3 +1,6 @@
+package interfaz;
+
+
 import java.awt.Color;
 import java.awt.EventQueue;
 import java.awt.GridLayout;
@@ -17,6 +20,16 @@ import javax.swing.UnsupportedLookAndFeelException;
 import javax.swing.border.TitledBorder;
 
 import com.itextpdf.text.DocumentException;
+
+import colecciones.Compania;
+import interfaz.agregar.FrameAgregarAdmin;
+import interfaz.agregar.FrameAgregarCliente;
+import interfaz.agregar.FrameAgregarOtroContrato;
+import interfaz.agregar.FrameAgregarPlan;
+import interfaz.eliminar.FrameEliminarCliente;
+import interfaz.eliminar.FrameEliminarContrato;
+import interfaz.modificar.FrameModificarCliente;
+import interfaz.reportar.FrameVerClientes;
 
 public class FrameInterfaz extends JFrame {
 
@@ -55,14 +68,14 @@ public class FrameInterfaz extends JFrame {
 		setResizable(false);
 		setTitle("MovilMaker");
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		setBounds(100, 100, 418, 466);
+		setBounds(100, 100, 630, 599);
 		contentPane = new JPanel();
 		contentPane.setBorder(new TitledBorder(UIManager.getBorder("TitledBorder.border"), "Bienvenido a",
 				TitledBorder.CENTER, TitledBorder.TOP, null, new Color(0, 0, 0)));
 		setContentPane(contentPane);
 		contentPane.setLayout(null);
 
-		// Creación del logo
+		// Creaciï¿½n del logo
 		JLabel label = new JLabel("");
 		label.setBounds(92, 24, 217, 54);
 		contentPane.add(label);
@@ -70,7 +83,7 @@ public class FrameInterfaz extends JFrame {
 		label.setIcon(new ImageIcon("resources\\logo.png"));
 
 		JPanel panel = new JPanel();
-		panel.setBounds(10, 89, 189, 244);
+		panel.setBounds(10, 89, 189, 326);
 		contentPane.add(panel);
 
 		JButton btnAgregarAdmin = new JButton("Agregar Administrador");
@@ -98,13 +111,30 @@ public class FrameInterfaz extends JFrame {
 				dispose();
 			}
 		});
-
-		JButton btnIngresarContratoclientes = new JButton("Ingresar contrato a cliente");
-		panel.add(btnIngresarContratoclientes);
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		
+		JButton btnAgregarPlan = new JButton("Agregar Plan");
+		btnAgregarPlan.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				FrameAgregarPlan fAgregarPlan = new FrameAgregarPlan(datosEmpresa);
+				fAgregarPlan.setVisible(true);
+				dispose();
+			}
+		});
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
-			btnIngresarContratoclientes.setEnabled(false);
-		btnIngresarContratoclientes.addActionListener(new ActionListener() {
+			btnAgregarPlan.setEnabled(false);
+		panel.add(btnAgregarPlan);
+		
+		JButton btnAgregarEquipo = new JButton("Agregar Equipo");
+		panel.add(btnAgregarEquipo);
+
+		JButton btnAgregarContratocliente = new JButton("Agregar contrato a cliente");
+		panel.add(btnAgregarContratocliente);
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
+		if (falta == 0 || falta == 1)
+			btnAgregarContratocliente.setEnabled(false);
+		btnAgregarContratocliente.addActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
 				FrameAgregarOtroContrato fOtroContrato = new FrameAgregarOtroContrato(datosEmpresa);
@@ -112,23 +142,9 @@ public class FrameInterfaz extends JFrame {
 				dispose();
 			}
 		});
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
 			btnAgregarClientes.setEnabled(false);
-
-		JButton btnActualizarDatosDe = new JButton("Actualizar datos de cliente");
-		panel.add(btnActualizarDatosDe);
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
-		if (falta == 0 || falta == 1)
-			btnActualizarDatosDe.setEnabled(false);
-		btnActualizarDatosDe.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				FrameActualizarCliente fActualizarCliente = new FrameActualizarCliente(datosEmpresa);
-				fActualizarCliente.setVisible(true);
-				dispose();
-			}
-		});
 
 		JPanel panel_1 = new JPanel();
 		panel_1.setBounds(209, 89, 189, 244);
@@ -137,7 +153,7 @@ public class FrameInterfaz extends JFrame {
 
 		JButton btnVerClientesActuales = new JButton("Ver Clientes Actuales");
 		panel_1.add(btnVerClientesActuales);
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
 			btnVerClientesActuales.setEnabled(false);
 		btnVerClientesActuales.addActionListener(new ActionListener() {
@@ -162,7 +178,7 @@ public class FrameInterfaz extends JFrame {
 
 		JButton btnTerminarContrato = new JButton("Terminar contrato");
 		panel_1.add(btnTerminarContrato);
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
 			btnTerminarContrato.setEnabled(false);
 		btnTerminarContrato.addActionListener(new ActionListener() {
@@ -173,19 +189,19 @@ public class FrameInterfaz extends JFrame {
 				dispose();
 			}
 		});
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
 			btnNewButton.setEnabled(false);
 
-		//BOTÓN QUE GENERA UN  ARCHIVO EN PDF DE LOS CLIENTES
+		//BOTï¿½N QUE GENERA UN  ARCHIVO EN PDF DE LOS CLIENTES
 		JButton btnGenerarReporte = new JButton("Generar Reporte PDF");
 		panel_1.add(btnGenerarReporte);
-		// Si no hay admin (Falta = 0 o 1) no se puede activar este botón
+		// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
 		if (falta == 0 || falta == 1)
 			btnGenerarReporte.setEnabled(false);
 		btnGenerarReporte.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				//BOTÓN QUE GENERA UN  ARCHIVO EN PDF DE LOS CLIENTES
+				//BOTï¿½N QUE GENERA UN  ARCHIVO EN PDF DE LOS CLIENTES
 				try {
 					datosEmpresa.reporte();
 					JOptionPane.showMessageDialog(null, "Reporte empresa creado", "Aviso", JOptionPane.INFORMATION_MESSAGE);
@@ -197,8 +213,28 @@ public class FrameInterfaz extends JFrame {
 		});
 
 		JButton btnSalir = new JButton("Salir");
-		btnSalir.setBounds(111, 344, 189, 61);
+		btnSalir.setBounds(209, 485, 189, 61);
 		contentPane.add(btnSalir);
+		
+		JPanel panel_2 = new JPanel();
+		panel_2.setBounds(425, 89, 189, 244);
+		contentPane.add(panel_2);
+		panel_2.setLayout(new GridLayout(0, 1, 0, 0));
+		
+				JButton btnActualizarDatosDe = new JButton("Actualizar datos de cliente");
+				panel_2.add(btnActualizarDatosDe);
+				btnActualizarDatosDe.addActionListener(new ActionListener() {
+					@Override
+					public void actionPerformed(ActionEvent e) {
+						FrameModificarCliente fActualizarCliente = new FrameModificarCliente(datosEmpresa);
+						fActualizarCliente.setVisible(true);
+						dispose();
+					}
+				});
+				// Si no hay admin (Falta = 0 o 1) no se puede activar este botï¿½n
+				if (falta == 0 || falta == 1)
+					btnActualizarDatosDe.setEnabled(false);
+				
 		btnSalir.addActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
