+++ src/interfaz/agregar/FrameAgregarCliente.java
@@ -1,3 +1,4 @@
+package interfaz.agregar;
 
 
 import java.awt.Color;
@@ -13,6 +14,12 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 import javax.swing.border.TitledBorder;
+
+import colecciones.Cliente;
+import colecciones.Compania;
+import extras.Database;
+import extras.XML;
+import interfaz.FrameInterfaz;
 
 import java.awt.*;
 import java.awt.event.ActionEvent;
