+++ src/interfaz/agregar/FrameAgregarOtroContrato.java
@@ -1,3 +1,4 @@
+package interfaz.agregar;
 
 
 import java.awt.Color;
@@ -11,6 +12,10 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 import javax.swing.border.TitledBorder;
+
+import colecciones.Cliente;
+import colecciones.Compania;
+import interfaz.FrameInterfaz;
 
 public class FrameAgregarOtroContrato extends JFrame {
